+++ dash_pydantic_form/__init__.py
@@ -5,6 +5,7 @@ from pydantic import BaseModel
 from dash_pydantic_form.fields import fields
 from dash_pydantic_form.form_section import FormSection, Sections
 from dash_pydantic_form.model_form import ModelForm
+from dash_pydantic_form.utils import from_form_data, get_model_cls
 
 BaseModel.__getitem__ = lambda self, key: self.__dict__.get(key)
 BaseModel.to_plotly_json = lambda self: self.model_dump(mode="json")
@@ -28,4 +29,4 @@ _js_dist = [
 
 __version__ = version(__package__)
 
-__all__ = ["FormSection", "ModelForm", "Sections", "fields", "__version__"]
+__all__ = ["FormSection", "ModelForm", "Sections", "get_model_cls", "fields", "from_form_data", "__version__"]